Add initial cut at User personas
This came up during Code for Boston in https://github.com/studentinsights/studentinsights/issues/208, and this is partially addressing https://github.com/studentinsights/studentinsights/issues/156.

The background here is that in working on improvements or features that involve some aspect of product or design, it's often helpful to talk about our users as a way to understand what problems they have and how we can help.  This is mostly done in-person, and we're not really capturing any of this in the project in a way that makes it easier for us as a diverse and often distributed team to build a shared understanding here.

Doing this in person is great! and we should keep doing it.  But this is a first step at capturing some of this as a way to help onboard new team members more easily, and as a way to make concrete some of the important background knowledge and assumptions that folks have who've been working on the project longer.